3496 - Move StatisticalGraphsPanel chart logic to a hook
