+++ src/client/pages/Geo/GeoMap/GeoMapMenuStatistics/StatisticalGraphsPanel/StatisticalGraphsPanel.scss
@@ -1,5 +1,5 @@
 @import 'src/client/style/partials';
 
-.statistical-graph-panel-container {
-  padding: 0 $spacing-s $spacing-s $spacing-s;
+.geo-statistical-graphs-panel__container {
+  width: 100%;
 }
